監視データint以上の値対応(Long型で扱うように変更)

+++ src/com/zabbix/MonitorActivity.java
@@ -175,7 +175,7 @@ public class MonitorActivity extends Activity {
         	TimeRange t = new TimeRange();
         	t.setTimeTill(historyDataList.get(i).getUnixtime());
         	if(item.getItemValueType().equals("3")) {
-        		series.add(new Second(t.getTimeTillAtDateType()),Integer.valueOf(historyDataList.get(i).getValue()));
+        		series.add(new Second(t.getTimeTillAtDateType()),Long.valueOf(historyDataList.get(i).getValue()));
         	}else if (item.getItemValueType().equals("0")) {
         		series.add(new Second(t.getTimeTillAtDateType()),Double.valueOf(historyDataList.get(i).getValue()));
         	}
@@ -285,7 +285,7 @@ public class MonitorActivity extends Activity {
         	TimeRange t = new TimeRange();
         	t.setTimeTill(historyDataList.get(i).getUnixtime());
         	if(item.getItemValueType().equals("3")) {
-        		series.add(new Second(t.getTimeTillAtDateType()),Integer.valueOf(historyDataList.get(i).getValue()));
+        		series.add(new Second(t.getTimeTillAtDateType()),Long.valueOf(historyDataList.get(i).getValue()));
         	}else if (item.getItemValueType().equals("0")) {
         		series.add(new Second(t.getTimeTillAtDateType()),Double.valueOf(historyDataList.get(i).getValue()));
         	}
